Fixes #13 which causes erroneous behavior on reconnect
When using the same groupDevice to reconnect, as would be the
case with Glass, which uses a static groupDevice, old websocket
would not be removed from channels_to_sockets, causing the new
one to be unregistered upon send to any channel where the old one
was subscribed. TODO: remove old subscriptions from
channels_to_groupDevices.